refactor(wallet): refactor xlm eth addresses
