+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/types.ts
@@ -19,8 +19,10 @@ type MessagesType = {
   'buttons.continue': 'Continue'
   'buttons.done': 'Done'
   'buttons.go_back': 'Go Back'
+  'buttons.hide': 'Hide'
   'buttons.learn_more': 'Learn More'
   'buttons.ok': 'OK'
+  'buttons.show': 'Show'
   'components.DateInputBox.placeholder.month': 'Month'
   'components.EmailVerification.changeemail': 'Change Email'
   'components.EmailVerification.checkinbox': 'Check your inbox. We sent an email to:'
@@ -1795,10 +1797,15 @@ type MessagesType = {
   'scenes.settings.addresses.btc.wallets.walletname': 'Wallet Name'
   'scenes.settings.addresses.delete_address': 'Delete'
   'scenes.settings.addresses.edit_name': 'Edit Label'
+  'scenes.settings.addresses.eth.address': 'Address Info'
+  'scenes.settings.addresses.eth.legacy': 'Legacy Address Info'
   'scenes.settings.addresses.menutop.bch': 'Bitcoin Cash'
   'scenes.settings.addresses.menutop.btc': 'Bitcoin'
   'scenes.settings.addresses.sign_message': 'Sign Message'
   'scenes.settings.addresses.unarchive': 'Unarchive'
+  'scenes.settings.addresses.viewkeys.header': 'View Your Private Keys.'
+  'scenes.settings.addresses.viewkeys.warning': 'Warning'
+  'scenes.settings.addresses.viewkeys.warning-body': 'Do not share your private keys with anyone. We will never ask for them. Sharing your keys may result in a loss of funds.'
   'scenes.settings.general.about.description': 'Learn more about our company.'
   'scenes.settings.general.about.title': 'About'
   'scenes.settings.general.explain': 'Use your Wallet ID to log in using our web client,'
